requirements.txt
Django==5.1.5
djangorestframework==3.16.0
gunicorn==23.0.0
psycopg2==2.9.10
dj_database_url==3.0.0